requirements.txt
plotly==5.22.0
openai==1.37.1
anthropic==0.42.0
google-ai-generativelanguage==0.6.10
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
google-pasta==0.2.0
googleapis-common-protos==1.66.0
astor==0.8.1
astroid==2.5.2
numpy==1.26.4
ipdb==0.13.13
streamlit==1.41.1
matplotlib
scipy
